requirements.txt
python-dotenv
google-generativeai
streamlit
transformers 
pillow 
torch
torchvision
ultralytics
accelerate